Swap Forgot link and Login button.
Seems kind of weird to have the Forgot link before the Login button. It also flows more naturally when using TAB to move focus around. I've been looking around and the trend seems to be Login, then, Forgot (Still no clear winner here).  This is a highly opinionated change proposal so feel free to dismiss it.